+++ apps/aklub/forms.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-
+from django import forms
 from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import ReadOnlyPasswordHashField, UserChangeForm, UserCreationForm
+from django.contrib.auth.forms import ReadOnlyPasswordHashField, UserChangeForm, UserCreationForm, UsernameField
 from django.shortcuts import reverse
 from django.urls import reverse_lazy
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
-from .models import UserProfile
+from .models import AdministrativeUnit, ProfileEmail
 from .views import get_unique_username
 
 Profile = get_user_model()
@@ -37,25 +37,30 @@ class UserCreateForm(UserCreationForm):
     def clean(self):
         if self.cleaned_data['email'] is None:
             return super().clean()
-        try:
-            user = UserProfile.objects.get(email=self.cleaned_data['email'])
-            url = reverse('admin:aklub_userprofile_change', args=(user.pk,))
+
+        email, created = ProfileEmail.objects.get_or_create(email=self.cleaned_data['email'])
+        if created is True:
+            return super().clean()
+
+        else:
+            url = reverse('admin:aklub_userprofile_change', args=(email.user.pk,))
             self.add_error(
                 'email',
                 mark_safe(
                     _(f'<a href="{url}">User with this email already exist in database, click here to edit</a>'),
                 ),
             )
-        except UserProfile.DoesNotExist:
-            return super().clean()
 
     def save(self, commit=True):
-        user = super(UserCreationForm, self).save(commit=False)
+        user = super().save(commit=False)
+
         user.username = self.cleaned_data['username']
         username_validation(user=user, fields=self.cleaned_data)
-
-        if commit:
-            user.save()
+        email = user.email
+        user.email = None
+        user.save()
+        if self.cleaned_data['email'] is not None:
+            ProfileEmail.objects.create(email=email, user=user, is_primary=True)
         return user
 
 
@@ -86,3 +91,83 @@ class UserUpdateForm(UserChangeForm):
         if commit:
             user.save()
         return user
+
+
+class UnitUserProfileAddForm(forms.ModelForm):
+    username = forms.CharField(required=False,)
+
+    class Meta:
+        model = Profile
+        fields = (
+            'username',
+            # 'first_name',
+            # 'last_name',
+            # 'title_before',
+            # 'title_after',
+            'email',
+            # 'useprofile__sex',
+            # 'birth_day',
+            # 'birth_month',
+            # 'age_group',
+            'note',
+            'administrative_units',
+            'street',
+            'city',
+            'country',
+            'zip_code',
+            'different_correspondence_address',
+            'correspondence_street',
+            'correspondence_city',
+            'correspondence_country',
+            'correspondence_zip_code',
+            'addressment',
+            'addressment_on_envelope',
+        )
+        field_classes = {'username': UsernameField}
+
+    def clean(self):
+        if self.cleaned_data['email'] is None:
+            return super().clean()
+
+        email, created = ProfileEmail.objects.get_or_create(email=self.cleaned_data['email'])
+        if created is True:
+            return super().clean()
+
+        else:
+            user = email.user
+            administrated_unit = AdministrativeUnit.objects.get(id=self.request.user.administrated_units.first().id)
+            user.administrative_units.add(administrated_unit)
+            url = reverse('admin:aklub_userprofile_change', args=(user.pk,))
+            self.add_error(
+                'email',
+                mark_safe(
+                    _(f'<a href="{url}">User with this email already exist in database and is available now, click here to edit</a>'),
+                ),
+            )
+
+    def save(self, commit=True):
+        user = super().save(commit=False)
+        email = user.email
+        user.email = None
+        user.save()
+        ProfileEmail.objects.create(email=email, user=user, is_primary=True)
+        return user
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['administrative_units'].queryset = self.request.user.administrated_units.all()
+        self.fields['administrative_units'].required = True
+
+
+class UnitUserProfileChangeForm(UnitUserProfileAddForm):
+
+    class Meta(UnitUserProfileAddForm.Meta):
+        pass
+
+    def clean(self):
+        pass
+
+    def __init__(self, *args, **kwargs):
+        super(UnitUserProfileAddForm, self).__init__(*args, **kwargs)
+        self.fields['administrative_units'].queryset = self.instance.administrative_units.all()
+        self.fields['administrative_units'].disabled = True
